# Remove more bycolumn loops