--- a/src/prognostic_equations/remaining_tendency.jl
+++ b/src/prognostic_equations/remaining_tendency.jl
@@ -52,13 +52,14 @@ function additional_tendency!(Yₜ, Y, p, t)
         edmfx_nh_pressure_tendency!(Yₜ, Y, p, t, colidx, p.turbconv_model)
         explicit_sgs_flux_tendency!(Yₜ, Y, p, t, colidx, p.turbconv_model)
         precipitation_tendency!(Yₜ, Y, p, t, colidx, p.precip_model)
-        # NOTE: All ρa tendencies should be applied before calling this function
-        pressure_work_tendency!(Yₜ, Y, p, t, colidx, p.atmos.turbconv_model)
-
-        # NOTE: This will zero out all monmentum tendencies in the edmfx advection test
-        # please DO NOT add additional velocity tendencies after this function
-        zero_velocity_tendency!(Yₜ, Y, p, t, colidx)
     end
+
+    # NOTE: All ρa tendencies should be applied before calling this function
+    pressure_work_tendency!(Yₜ, Y, p, t, p.atmos.turbconv_model)
+
+    # NOTE: This will zero out all monmentum tendencies in the edmfx advection test
+    # please DO NOT add additional velocity tendencies after this function
+    zero_velocity_tendency!(Yₜ, Y, p, t)
     # TODO: make bycolumn-able
     non_orographic_gravity_wave_tendency!(
         Yₜ,