This PR removes more `bycolumn` loops. These calls are inside the implicit tendency, which are still called with explicit solvers (except that the tendencies are added explicitly). This was found in this [build](https://buildkite.com/clima/climaatmos-ci/builds/10597#0188d9a4-752f-4a25-8841-b4d7aebff0bc).

This should fix some rather large performance issues for the gpu example (cc @bloops)

The benchmark is looking more promising:

Main branch:
`[ Info: Ran step! 10 times in 17.764 s (45437150 allocations: 9.56 GiB), (1.780 s per step)`
This PR:
`[ Info: Ran step! 10 times in 127.031 ms (244420 allocations: 35.30 MiB), (15.710 ms per step)`

I will say, one thing is very odd: why is the benchmark, which runs 20 steps, so much faster than the target job, which runs 60 steps? (at `15.710 ms` per step, the difference should be < 1 second)