@simonbyrne 2023-06-21
was this on the CPU or GPU?