@charleskawczynski 2023-06-21
> was this on the CPU or GPU?

This was for the gpu example