@simonbyrne 2023-06-21
It could just be compilation time? GPU kernels are much more expensive to compile.