@bloops 2023-06-21
> I will say, one thing is very odd: why is the benchmark, which runs 20 steps, so much faster than the target job, which runs 60 steps? (at 15.710 ms per step, the difference should be < 1 second)

I saw that the first 10 steps (which includes compilation time) in the benchmark takes 40x more walltime than the next 10 steps (even without this PR). So perhaps the target job running 60 steps is also counting the compilation time?