@charleskawczynski 2023-06-21
Ah, my mistake. We modify `h_elem` in the buildkite performance job (I did it originally kind of as a demo). Let's just remove that argument. That should make it more similar.