@charleskawczynski 2023-06-22
@bloops, we're going to close this in favor of https://github.com/CliMA/ClimaCore.jl/pull/1341 and https://github.com/CliMA/ClimaCore.jl/pull/1332.